move enter_debugger
