+++ spyder_kernels/customize/spydercustomize.py
@@ -27,7 +27,7 @@ from IPython.core.getipython import get_ipython
 
 from spyder_kernels.comms.frontendcomm import CommError, frontend_request
 from spyder_kernels.customize.namespace_manager import NamespaceManager
-from spyder_kernels.customize.spyderpdb import SpyderPdb
+from spyder_kernels.customize.spyderpdb import SpyderPdb, enter_debugger
 from spyder_kernels.customize.umr import UserModuleReloader
 from spyder_kernels.py3compat import TimeoutError, PY2, _print, encode
 
@@ -571,54 +571,6 @@ def runfile(filename=None, args=None, wdir=None, namespace=None,
 builtins.runfile = runfile
 
 
-def enter_debugger(filename, continue_if_has_breakpoints, code_format):
-    """Enter debugger. Code format should be a format that accept filename."""
-    if filename is None:
-        filename = get_current_file_name()
-        if filename is None:
-            return
-
-    kernel = get_ipython().kernel
-    recursive = kernel.is_debugging()
-    if recursive:
-        parent_debugger = kernel._pdb_obj
-        sys.settrace(None)
-        globals = parent_debugger.curframe.f_globals
-        locals = parent_debugger.curframe_locals
-        # Create child debugger
-        debugger = parent_debugger.__class__(
-            completekey=parent_debugger.completekey,
-            stdin=parent_debugger.stdin, stdout=parent_debugger.stdout)
-        debugger.use_rawinput = parent_debugger.use_rawinput
-        debugger.prompt = "(%s) " % parent_debugger.prompt.strip()
-    else:
-        debugger = pdb.Pdb()
-
-    filename = debugger.canonic(filename)
-    debugger._wait_for_mainpyfile = True
-    debugger.mainpyfile = filename
-    debugger.continue_if_has_breakpoints = continue_if_has_breakpoints
-    debugger._user_requested_quit = False
-
-    if os.name == 'nt':
-        filename = filename.replace('\\', '/')
-
-    code = code_format.format(repr(filename))
-
-    if recursive:
-        # Enter recursive debugger
-        parent_debugger.message("ENTERING RECURSIVE DEBUGGER")
-        sys.call_tracing(debugger.run, (code, globals, locals))
-        parent_debugger.message("LEAVING RECURSIVE DEBUGGER")
-        # Reset parent debugger
-        sys.settrace(parent_debugger.trace_dispatch)
-        parent_debugger.lastcmd = debugger.lastcmd
-        kernel._register_pdb_session(parent_debugger)
-    else:
-        # The breakpoint might not be in the cell
-        debugger.run(code)
-
-
 def debugfile(filename=None, args=None, wdir=None, post_mortem=False,
               current_namespace=False):
     """
@@ -629,6 +581,11 @@ def debugfile(filename=None, args=None, wdir=None, post_mortem=False,
     """
     # Tell IPython to hide this frame (>7.16)
     __tracebackhide__ = True
+    if filename is None:
+        filename = get_current_file_name()
+        if filename is None:
+            return
+
     enter_debugger(
         filename, True,
         "runfile({}" +
@@ -706,6 +663,11 @@ def debugcell(cellname, filename=None, post_mortem=False):
     """Debug a cell."""
     # Tell IPython to hide this frame (>7.16)
     __tracebackhide__ = True
+    if filename is None:
+        filename = get_current_file_name()
+        if filename is None:
+            return
+
     enter_debugger(
         filename, False,
         "runcell({}, ".format(repr(cellname)) +
