+++ spyder_kernels/customize/spyderpdb.py
@@ -7,8 +7,9 @@
 """Spyder debugger."""
 
 import bdb
-import sys
 import logging
+import os
+import sys
 import traceback
 from collections import namedtuple
 
@@ -700,3 +701,44 @@ class SpyderPdb(ipyPdb, object):  # Inherits `object` to call super() in PY2
         """
         with DebugWrapper(self):
             super(SpyderPdb, self).runcall(*args, **kwds)
+
+
+def enter_debugger(filename, continue_if_has_breakpoints, code_format):
+    """Enter debugger. Code format should be a format that accept filename."""
+    kernel = get_ipython().kernel
+    recursive = kernel.is_debugging()
+    if recursive:
+        parent_debugger = kernel._pdb_obj
+        sys.settrace(None)
+        globals = parent_debugger.curframe.f_globals
+        locals = parent_debugger.curframe_locals
+        # Create child debugger
+        debugger = SpyderPdb(
+            completekey=parent_debugger.completekey,
+            stdin=parent_debugger.stdin, stdout=parent_debugger.stdout)
+        debugger.use_rawinput = parent_debugger.use_rawinput
+        debugger.prompt = "(%s) " % parent_debugger.prompt.strip()
+    else:
+        debugger = SpyderPdb()
+
+    filename = debugger.canonic(filename)
+    debugger._wait_for_mainpyfile = True
+    debugger.mainpyfile = filename
+    debugger.continue_if_has_breakpoints = continue_if_has_breakpoints
+    debugger._user_requested_quit = False
+
+    if os.name == 'nt':
+        filename = filename.replace('\\', '/')
+
+    code = code_format.format(repr(filename))
+
+    if recursive:
+        # Enter recursive debugger
+        sys.call_tracing(debugger.run, (code, globals, locals))
+        # Reset parent debugger
+        sys.settrace(parent_debugger.trace_dispatch)
+        parent_debugger.lastcmd = debugger.lastcmd
+        kernel._register_pdb_session(parent_debugger)
+    else:
+        # The breakpoint might not be in the cell
+        debugger.run(code)
